fixbug：修复发送模板消息结果中的msgid实际返回结果是long类型报错的问题

+++ Senparc.Weixin.MP/Senparc.Weixin.MP/AdvancedAPIs/TemplateMessage/SendTemplateMessageResult.cs
@@ -14,6 +14,6 @@ namespace Senparc.Weixin.MP.AdvancedAPIs
         /// <summary>
         /// msgid
         /// </summary>
-        public int msgid { get; set; }
+        public long msgid { get; set; }
     }
 }
